Extract handler classes into pieces
