+++ monitor_app/views_collection/handlers/ModelDataHandler.py
@@ -1,0 +1,8 @@
+import abc
+class ModelDataHandler(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def getData(self):
+        return NotImplemented
+    @abc.abstractmethod
+    def getTitle(self):
+        return NotImplemented
