+++ monitor_app/views_collection/handlers/__pycache__/.gitignore
@@ -1,0 +1,1 @@
+*.pyc
